requirements.txt
Django==3.2.16
djangorestframework==3.14.0
PyMySQL==0.9.3
gunicorn==19.9.0
mysqlclient==2.0.3
